Also install amplify-cli for Node.js 12
With this change it is possible to use Node.js 12 throughout the whole amplify build process.
Since node 10 is set as the default node version it doesnt seem necessary to install amplify-cli for the default node version and node 10, so this commit installs the amplify-cli for node 12 instead of node 10.